+++ example/android/app/src/main/kotlin/com/example/imin_cash_drawer_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.imin_cash_drawer_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}