+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
 
-* TODO: Describe initial release.
+# Changelog
+
+## 0.0.1 (2024-10-03)
+- Initial release of the `imin_cash_drawer` package.
+- Added functionality to check cash drawer status.
+- Implemented method to open the cash drawer connected to IMIN POS devices.
+
+